[EI-241] [api] test: test refactoring
