+++ api/src/numerical-guidance/test/integration-test/adapter/persistence/indicator-board-metadata.persistent.adapter.spec.ts
@@ -20,10 +20,10 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
   let indicatorBoardMetadataPersistentAdapter: IndicatorBoardMetadataPersistentAdapter;
   const seeding = async () => {
     const memberRepository = dataSource.getRepository(MemberEntity);
-    await memberRepository.insert({ id: '10' });
-    await memberRepository.insert({ id: '5' });
-    await memberRepository.insert({ id: '999' });
-    await memberRepository.insert({ id: '9999' });
+    await memberRepository.insert({ id: '10', email: 'test@gmail.com' });
+    await memberRepository.insert({ id: '5', email: 'test@gmail.com' });
+    await memberRepository.insert({ id: '999', email: 'test@gmail.com' });
+    await memberRepository.insert({ id: '9999', email: 'test@gmail.com' });
     memberRepository.save;
 
     const indicatorBoardMetadataRepository = dataSource.getRepository(IndicatorBoardMetadataEntity);
@@ -33,7 +33,7 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
       indicatorInfos: [],
       customForecastIndicatorIds: [],
       sections: { section1: [] },
-      member: { id: '10' },
+      member: { id: '10', email: 'test@gmail.com' },
     });
 
     await indicatorBoardMetadataRepository.insert({
@@ -42,7 +42,7 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
       indicatorInfos: [],
       customForecastIndicatorIds: [],
       sections: { section1: [] },
-      member: { id: '5' },
+      member: { id: '5', email: 'test@gmail.com' },
     });
 
     await indicatorBoardMetadataRepository.insert({
@@ -51,7 +51,7 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
       indicatorInfos: [],
       customForecastIndicatorIds: [],
       sections: { section1: [] },
-      member: { id: '5' },
+      member: { id: '5', email: 'test@gmail.com' },
     });
 
     await indicatorBoardMetadataRepository.insert({
@@ -60,7 +60,7 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
       indicatorInfos: [],
       customForecastIndicatorIds: [],
       sections: { section1: [] },
-      member: { id: '5' },
+      member: { id: '5', email: 'test@gmail.com' },
     });
 
     await indicatorBoardMetadataRepository.insert({
@@ -86,7 +86,7 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
       sections: {
         section1: ['indicatorId1', 'indicatorId2', 'customForecastIndicatorId1', 'customForecastIndicatorId2'],
       },
-      member: { id: '999' },
+      member: { id: '999', email: 'test@gmail.com' },
     });
 
     await indicatorBoardMetadataRepository.insert({
@@ -112,7 +112,7 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
       sections: {
         section1: ['indicatorId1', 'indicatorId2', 'customForecastIndicatorId1', 'customForecastIndicatorId2'],
       },
-      member: { id: '999' },
+      member: { id: '999', email: 'test@gmail.com' },
     });
 
     await indicatorBoardMetadataRepository.insert({
@@ -138,7 +138,7 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
       sections: {
         section1: ['indicatorId1', 'indicatorId2', 'customForecastIndicatorId1', 'customForecastIndicatorId2'],
       },
-      member: { id: '999' },
+      member: { id: '999', email: 'test@gmail.com' },
     });
 
     await indicatorBoardMetadataRepository.insert({
@@ -147,7 +147,7 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
       indicatorInfos: [],
       customForecastIndicatorIds: [],
       sections: { section1: [] },
-      member: { id: '9999' },
+      member: { id: '9999', email: 'test@gmail.com' },
     });
 
     await indicatorBoardMetadataRepository.insert({
@@ -183,7 +183,7 @@ describe('IndicatorBoardMetadataPersistentAdapter', () => {
           'customForecastIndicatorId3',
         ],
       },
-      member: { id: '9999' },
+      member: { id: '9999', email: 'test@gmail.com' },
     });
   };
