+++ api/src/numerical-guidance/test/unit-test/query/indicator-board-metadata/get-indicator-board-metadata.query.handler.spec.ts
@@ -21,10 +21,9 @@ describe('GetIndicatorBoardMetadataQueryHandler', () => {
 
   const seeding = async () => {
     const memberRepository = dataSource.getRepository(MemberEntity);
-    await memberRepository.insert({ id: '10' });
-    await memberRepository.insert({ id: '5' });
-    await memberRepository.insert({ id: '999' });
-    memberRepository.save;
+    await memberRepository.insert({ id: '10', email: 'test@gmail.com' });
+    await memberRepository.insert({ id: '5', email: 'test@gmail.com' });
+    await memberRepository.insert({ id: '999', email: 'test@gmail.com' });
 
     const indicatorBoardMetadataRepository = dataSource.getRepository(IndicatorBoardMetadataEntity);
     await indicatorBoardMetadataRepository.insert({
@@ -33,7 +32,7 @@ describe('GetIndicatorBoardMetadataQueryHandler', () => {
       indicatorInfos: [],
       customForecastIndicatorIds: [],
       sections: { section1: [] },
-      member: { id: '10' },
+      member: { id: '10', email: 'test@gmail.com' },
     });
   };
